+++ weByteProject/budgetView.swift
@@ -11,10 +11,10 @@ import SlidingTabView
 struct budgetView: View {
     
     @State private var tabIndex=0
-    var incomeText = 0
-    @Binding var goalText : Int
-    @Binding var spendingText :Int
-    @Binding var saving :Int
+    var incomeText = 0.0
+    @Binding var goalText : Double
+    @Binding var spendingText :Double
+    @Binding var saving :Double
     @Binding var stringGoal: String
     
     var body: some View {
@@ -37,7 +37,7 @@ struct budgetView: View {
                 SlidingTabView(selection: $tabIndex, tabs: ["الميزانية","خريطة الطريق"], animation:.easeOut)
                 Spacer().frame(height:50)
                 if tabIndex == 0{
-                    budgetCradView(incomeText: Int(incomeText) ,goalText: $goalText  ,spendingText: $spendingText, saving:$saving, stringGoal: $stringGoal )
+                    budgetCradView(incomeText: incomeText ,goalText: $goalText  ,spendingText: $spendingText, saving:$saving, stringGoal: $stringGoal )
                 }else if tabIndex == 1{
                     roadmapView()
                 }
@@ -55,9 +55,9 @@ struct budgetView: View {
 }
 
 struct budgetView_Previews: PreviewProvider {
-    @State static private var dummySpendingText = 0
-    @State static private var  dummysaving = 0
-    @State static private var  dummygoal = 0
+    @State static private var dummySpendingText = 0.0
+    @State static private var  dummysaving = 0.0
+    @State static private var  dummygoal = 0.0
     @State static private var  dummystringgoal = ""
